requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-pubsub>=2.18.0
google-cloud-secret-manager>=2.18.0
firebase-admin>=6.0.0
google-cloud-dialogflow-cx>=1.20.0
requests>=2.31.0

